+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Added
+- Added new separate services page for managing them as bookings.
+- Added service editor in the modal.
+
+### Removed
+- Removed the partial metabox application for editing and creating a service.
+
 ## [0.2.3] - 2018-10-08
 ### Added
 - Wizard labels editor.